split widget file into separate components
