+++ web/src/routes/_authorized/_default.tsx
@@ -1,6 +1,6 @@
 import { createFileRoute, Outlet } from '@tanstack/react-router';
 import { Navigation } from '../../shared/components/Navigation/Navigation';
-import { VideoSupportWidget } from '../../shared/video-support/widget';
+import { VideoSupportWidget } from '../../shared/video-support/VideoSupportWidget';
 
 export const Route = createFileRoute('/_authorized/_default')({
   component: RouteComponent,
