+++ web/src/shared/video-support/components/Thumbnail/style.scss
@@ -1,0 +1,26 @@
+.video-support-thumbnail {
+  flex-shrink: 0;
+  width: 106px;
+  height: 60px;
+  border-radius: var(--radius-md);
+  overflow: hidden;
+  background-color: var(--bg-emphasis);
+  position: relative;
+
+  img {
+    width: 100%;
+    height: 100%;
+    object-fit: cover;
+    display: none;
+
+    &.loaded {
+      display: block;
+    }
+  }
+
+  &.placeholder {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+  }
+}
