+++ web/src/shared/video-support/components/VideoOverlay/style.scss
@@ -1,0 +1,43 @@
+// ---------------------------------------------------------------------------
+// Video modal
+// ModalFoundation provides the backdrop and portal; we style the content here.
+// ---------------------------------------------------------------------------
+
+.video-support-modal-container {
+  position: relative;
+  display: inline-block;
+  padding-top: var(--size-2xl);
+}
+
+.video-support-modal {
+  background: var(--bg-default);
+  border-radius: var(--radius-lg);
+  box-shadow: 0 6px 30px 0 rgb(0 0 0 / 9%);
+  padding: var(--spacing-xl);
+  width: 920px;
+  max-width: 90vw;
+
+  iframe {
+    display: block;
+    width: 100%;
+    aspect-ratio: 16 / 9;
+    border: none;
+    border-radius: var(--radius-sm);
+  }
+}
+
+// Modal close button — floats outside the modal card (top-right).
+// Uses IconButton; overrides border-radius token to circular and adds background/shadow.
+
+.video-support-modal-close {
+  --button-border-radius-sm: var(--radius-full);
+
+  position: absolute;
+  top: var(--spacing-xs);
+  right: calc(-1 * var(--size-2xl));
+  background-color: var(--bg-default);
+  box-shadow: 0 2px 8px 0 rgb(0 0 0 / 10%);
+  z-index: 1;
+
+  @include animate(box-shadow);
+}
